refactor(portal-frontend): SPEC-PORTAL-TAXONOMY-SPLIT-001 commit 2 — extract _components/TagCloud.tsx
Pure renderer extraction. TagCloud has no state, no side-effects, no
mutations. Same prop signature as the inline version.

No test added — there's no logic to characterise. Behavior is locked
by the JSX itself; future regressions would be visible to the Playwright
pass on Voys (AC6).

Files:
- new: _components/TagCloud.tsx (51 lines)
- modified: _components/TaxonomyTab.tsx — 937 → 893 lines (-44)
  - removed inline TagCloud function (regels 340-381 in the post-commit-1 file)
  - added: import { TagCloud } from './TagCloud'

Gates:
- tsc -b --force: clean
- npx eslint on changed files: clean
- vitest run: 278/278 (unchanged)
