+++ klai-portal/frontend/src/routes/app/knowledge/$kbSlug/_components/TaxonomyTab.tsx
@@ -45,6 +45,7 @@ import {
   useTopTags,
   type SuggestState,
 } from '../-taxonomy-hooks'
+import { TagCloud } from './TagCloud'
 
 // SPEC-TAXONOMY-REVIEW-FLOW-001 follow-up: cap on healthy taxonomy size.
 // Mirrors the backend's ``taxonomy_consolidate_target_max`` (default 9).
@@ -333,49 +334,6 @@ function CoverageWidget({
           </span>
         </div>
       )}
-    </div>
-  )
-}
-
-// -- Tag cloud ----------------------------------------------------------------
-
-function TagCloud({
-  tags,
-  activeTags,
-  onTagClick,
-}: {
-  tags: { tag: string; count: number }[]
-  activeTags: Set<string>
-  onTagClick: (tag: string) => void
-}) {
-  const maxCount = tags[0]?.count ?? 1
-
-  return (
-    <div className="flex flex-wrap gap-1.5">
-      {tags.map(({ tag, count }) => {
-        const isActive = activeTags.has(tag)
-        // Scale font size from 0.75rem (min count) to 1rem (max count)
-        const scale = maxCount > 1 ? (count - 1) / (maxCount - 1) : 0
-        const fontSize = 0.75 + scale * 0.25
-
-        return (
-          <button
-            key={tag}
-            type="button"
-            onClick={() => onTagClick(tag)}
-            className={[
-              'inline-flex items-center gap-1 rounded-full border px-2.5 py-0.5 transition-colors',
-              isActive
-                ? 'border-gray-900 bg-gray-900 text-white'
-                : 'border-gray-200 bg-gray-50 text-gray-900 hover:bg-gray-100',
-            ].join(' ')}
-            style={{ fontSize: `${fontSize}rem` }}
-          >
-            <span>{tag}</span>
-            <span className="text-xs opacity-60 tabular-nums">{count}</span>
-          </button>
-        )
-      })}
     </div>
   )
 }
